@@ -2,6 +2,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.IO;
 using System.Linq;
 using System.Text;
@@ -38,13 +39,50 @@ namespace UndertaleModTool
         public UndertaleSoundEditor()
         {
             InitializeComponent();
-            this.Unloaded += Unload;
+            Unloaded += OnUnloaded;
+            DataContextChanged += OnDataContextChanged;
         }
 
-        public void Unload(object sender, RoutedEventArgs e)
+        private void OnUnloaded(object sender, RoutedEventArgs e)
         {
-            if (waveOut != null)
-                waveOut.Stop();
+            waveOut?.Stop();
+
+            if (DataContext is UndertaleSound oldObj)
+            {
+                oldObj.PropertyChanged -= OnPropertyChanged;
+            }
+        }
+
+        private void OnDataContextChanged(object sender, DependencyPropertyChangedEventArgs e)
+        {
+            if (e.OldValue is UndertaleSound oldObj)
+            {
+                oldObj.PropertyChanged -= OnPropertyChanged;
+            }
+            if (e.NewValue is UndertaleSound newObj)
+            {
+                newObj.PropertyChanged += OnPropertyChanged;
+            }
+        }
+
+        private void OnPropertyChanged(object sender, PropertyChangedEventArgs e)
+        {
+            OnAssetUpdated();
+        }
+
+        private void OnAssetUpdated()
+        {
+            if (mainWindow.Project is null || !mainWindow.IsSelectedProjectExportable)
+            {
+                return;
+            }
+            Dispatcher.BeginInvoke(() =>
+            {
+                if (DataContext is UndertaleSound obj)
+                {
+                    mainWindow.Project?.MarkAssetForExport(obj);
+                }
+            });
         }
 
         private void InitAudio()
